+++ package.json
@@ -10,6 +10,7 @@
     "express": "^4.21.1",
     "express-session": "^1.18.1",
     "file-type": "^16.5.4",
+    "form-data": "^4.0.1",
     "image-type": "^5.2.0",
     "jimp": "^1.6.0",
     "jsonwebtoken": "^9.0.2",
